+++ kala-collection/src/main/java/kala/collection/SeqView.java
@@ -18,7 +18,10 @@ package kala.collection;
 import kala.annotations.DelegateBy;
 import kala.collection.immutable.ImmutableArray;
 import kala.collection.internal.convert.AsJavaConvert;
+import kala.function.CheckedConsumer;
+import kala.function.CheckedIndexedConsumer;
 import kala.function.IndexedBiConsumer;
+import kala.function.IndexedConsumer;
 import kala.function.IndexedFunction;
 import kala.function.Predicates;
 import kala.index.Index;
@@ -28,6 +31,7 @@ import kala.tuple.Tuple2;
 import kala.collection.internal.view.SeqViews;
 import kala.annotations.Covariant;
 import kala.collection.internal.CollectionHelper;
+import org.jetbrains.annotations.ApiStatus;
 import org.jetbrains.annotations.Contract;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
@@ -307,5 +311,55 @@ public interface SeqView<@Covariant E> extends CollectionView<E>, SeqLike<E>, An
             set.add(e);
         }
         return ImmutableArray.Unsafe.<E>wrap(set.toArray()).view();
+    }
+
+    @Override
+    @ApiStatus.NonExtendable
+    @DelegateBy("forEach(Consumer<E, Ex>)")
+    @Contract(value = "_ -> this", pure = true)
+    default @NotNull SeqView<E> onEach(@NotNull Consumer<? super E> action) {
+        forEach(action);
+        return this;
+    }
+
+    @Override
+    @ApiStatus.NonExtendable
+    @DelegateBy("onEach(Consumer<E, Ex>)")
+    @Contract(value = "_ -> this", pure = true)
+    default <Ex extends Throwable> @NotNull SeqView<E> onEachChecked(@NotNull CheckedConsumer<? super E, ? extends Ex> action) throws Ex {
+        return onEach(action);
+    }
+
+    @Override
+    @ApiStatus.NonExtendable
+    @DelegateBy("onEach(Consumer<T, Ex>)")
+    @Contract(value = "_ -> this", pure = true)
+    default @NotNull SeqView<E> onEachUnchecked(@NotNull CheckedConsumer<? super E, ?> action) {
+        return onEach(action);
+    }
+
+    @Override
+    @ApiStatus.NonExtendable
+    @DelegateBy("forEachIndexed(IndexedConsumer<T, Ex>)")
+    @Contract(value = "_ -> this", pure = true)
+    default @NotNull SeqView<E> onEachIndexed(@NotNull IndexedConsumer<? super E> action) {
+        forEachIndexed(action);
+        return this;
+    }
+
+    @Override
+    @ApiStatus.NonExtendable
+    @DelegateBy("onEachIndexed(IndexedConsumer<E, Ex>)")
+    @Contract(value = "_ -> this", pure = true)
+    default <Ex extends Throwable> @NotNull SeqView<E> onEachChecked(@NotNull CheckedIndexedConsumer<? super E, ? extends Ex> action) throws Ex {
+        return onEachIndexed(action);
+    }
+
+    @Override
+    @ApiStatus.NonExtendable
+    @DelegateBy("onEachIndexed(IndexedConsumer<E, Ex>)")
+    @Contract(value = "_ -> this", pure = true)
+    default @NotNull SeqView<E> onEachUnchecked(@NotNull CheckedIndexedConsumer<? super E, ?> action) {
+        return onEachIndexed(action);
     }
 }